requirements.txt
# --- 后端框架 ---
fastapi[all]          # FastAPI 核心框架（包含 pydantic 和 email-validator）
uvicorn[standard]     # 高性能 ASGI 服务器

# --- 数据库驱动 ---
pymysql               # MySQL 数据库连接驱动

# --- 安全与加密 ---
pyjwt                 # JWT Token 签发与解析
passlib[bcrypt]       # 工业级密码哈希处理
cryptography>=3.4.7

# --- 自动化测试 ---
pytest                # 单元与集成测试框架
requests              # 测试脚本发送 API 请求
allure-pytest         # 生成 Allure 专业测试报告
locust
pytest-mock

# --- 环境配置 ---
python-dotenv         # 自动从 .env 文件加载环境变量

pytest-cov            # 代码覆盖率统计插件

alembic               # 数据库版本迁移与控制工具

redis                 # 高性能键值存储驱动